headshot tracking
the script makes a table of all currently spawned units and keeps tracks of the last time their head region was damaged, if it corresponds to tick the unit died in it then counts as a headshot (snapshot medal used as placeholder).
the issue is that not all units in the base game have a separate head region (grunts for example) so a workaround would need to be found for this case.
I also noticed a slight performance hit probably caused by the instance table.
not very proud of this one but this'll do for now